+++ riverdriver/riversqlite/river_sqlite_driver.go
@@ -1099,28 +1099,29 @@ func (e *Executor) JobUpdateFull(ctx context.Context, params *riverdriver.JobUpd
 	return jobRowFromInternal(job)
 }
 
-func (e *Executor) LeaderAttemptElect(ctx context.Context, params *riverdriver.LeaderElectParams) (bool, error) {
-	numElectionsWon, err := dbsqlc.New().LeaderAttemptElect(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderAttemptElectParams{
+func (e *Executor) LeaderAttemptElect(ctx context.Context, params *riverdriver.LeaderElectParams) (*riverdriver.Leader, error) {
+	leader, err := dbsqlc.New().LeaderAttemptElect(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderAttemptElectParams{
 		LeaderID: params.LeaderID,
 		Now:      timeStringNullable(params.Now),
 		TTL:      durationAsString(params.TTL),
 	})
 	if err != nil {
-		return false, interpretError(err)
+		return nil, interpretError(err)
 	}
-	return numElectionsWon > 0, nil
+	return leaderFromInternal(leader), nil
 }
 
-func (e *Executor) LeaderAttemptReelect(ctx context.Context, params *riverdriver.LeaderElectParams) (bool, error) {
-	numElectionsWon, err := dbsqlc.New().LeaderAttemptReelect(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderAttemptReelectParams{
-		LeaderID: params.LeaderID,
-		Now:      timeStringNullable(params.Now),
-		TTL:      durationAsString(params.TTL),
+func (e *Executor) LeaderAttemptReelect(ctx context.Context, params *riverdriver.LeaderReelectParams) (*riverdriver.Leader, error) {
+	leader, err := dbsqlc.New().LeaderAttemptReelect(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderAttemptReelectParams{
+		ElectedAt: timeString(params.ElectedAt),
+		LeaderID:  params.LeaderID,
+		Now:       timeStringNullable(params.Now),
+		TTL:       durationAsString(params.TTL),
 	})
 	if err != nil {
-		return false, interpretError(err)
+		return nil, interpretError(err)
 	}
-	return numElectionsWon > 0, nil
+	return leaderFromInternal(leader), nil
 }
 
 func (e *Executor) LeaderDeleteExpired(ctx context.Context, params *riverdriver.LeaderDeleteExpiredParams) (int, error) {
@@ -1154,7 +1155,10 @@ func (e *Executor) LeaderInsert(ctx context.Context, params *riverdriver.LeaderI
 }
 
 func (e *Executor) LeaderResign(ctx context.Context, params *riverdriver.LeaderResignParams) (bool, error) {
-	numResigned, err := dbsqlc.New().LeaderResign(schemaTemplateParam(ctx, params.Schema), e.dbtx, params.LeaderID)
+	numResigned, err := dbsqlc.New().LeaderResign(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderResignParams{
+		ElectedAt: timeString(params.ElectedAt),
+		LeaderID:  params.LeaderID,
+	})
 	if err != nil {
 		return false, interpretError(err)
 	}